refactor : 중복 검사 이후 코드 queries로 이동
